Add ulimit -s 16384 to avoid segfault on PPC64 during make check
The issue is related to the stack being too small when running the
ok-errors.R test case.  This test will cause an infinite recursion
loop, which should be interrupted by R.  On x86_64, this works as
expected.  On ppc64, the stack size needs to be doubled to 16384
(ulimit -s 16384) to be able to run successfully, otherwise the
segfault will happen.

RHBZ: 1136388

Signed-off-by: David Sommerseth <davids@redhat.com>